requirements.txt
# Core dependencies
requests>=2.28.0
numpy>=1.21.0
PyYAML>=6.0
tqdm>=4.60.0
joblib>=1.1.0
